requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==3.0.1
constantly==15.1.0
cryptography==39.0.1
daphne==4.0.0
defusedxml==0.7.1
Django==4.1.7
django-allauth==0.52.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
msgpack==1.0.4
oauthlib==3.2.2
Pillow==9.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.0.0
python-dotenv==1.0.0
python3-openid==3.2.0
redis==4.5.1
requests==2.28.2
requests-oauthlib==1.3.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
stripe==5.4.0
Twisted==22.10.0
txaio==23.1.1
typing-extensions==4.5.0
urllib3==1.26.14
zope.interface==5.5.2
